requirements.txt
fastapi
uvicorn
paho-mqtt
pydantic
sqlalchemy
python-multipart
